backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-storages
boto3
pillow
